Modify Move tests to jsonplaceholderTests package
