+++ src/test/java/jsonplaceholderTests/BaseTest.java
@@ -1,3 +1,5 @@
+package jsonplaceholderTests;
+
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.builder.ResponseSpecBuilder;
