+++ src/test/java/jsonplaceholderTests/DeletePost.java
@@ -1,3 +1,5 @@
+package jsonplaceholderTests;
+
 import org.junit.jupiter.api.Test;
 
 import static io.restassured.RestAssured.given;
